broke down lib/db in to files
